Moved calendar to its own component file, fixed appearence
